Added feature to export images of the board
Frame and Board now has two extra attributes: boolean record and int frameNumber. One can specify that he wants the screen to be captured to a number of frames. The images generated would then be stored in the output folder. FFMPEG can then be used to generate a mp4 file that can be opened using VLC